# apps/api/public/openapi.yaml
openapi: 3.0.3
info:
  title: sulfur database api
  contact:
    email: Public.kyuuanago@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000/
tags:
  - name: item
    description: Itemの操作
  - name: material
    description: Materialの操作
  - name: recipe
    description: Recipeの操作
paths:
  /item:
    get:
      tags:
        - item
      summary: Itemの取得
      description: nameで昇順ソートしてます
      operationId: getItem
      parameters:
        - name: nameLike
          in: query
          description: nameのlike検索
          required: false
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemArray"
              examples:
                ItemArray:
                  $ref: "#/components/examples/ItemArrayExample"
    post:
      tags:
        - item
      summary: Itemの新規作成
      description: ""
      operationId: createItem
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: integer
                tradeInPrice:
                  type: integer
                size:
                  type: integer
                  minimum: 1
                healAmount:
                  type: integer
                healSecond:
                  type: integer

      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        400:
          description: fail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: fail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /material:
    get:
      tags:
        - material
      summary: Materialの取得
      description: item.nameで昇順ソートしてます
      operationId: getMaterial
      parameters:
        - name: itemNameLike
          in: query
          description: item.nameのlike検索
          required: false
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialArray"
              examples:
                MaterialArray:
                  $ref: "#/components/examples/MaterialArrayExample"
  /recipe:
    get:
      tags:
        - recipe
      summary: Recipeの取得
      description: result.nameで昇順ソートしてます
      operationId: getRecipe
      parameters:
        - name: resultNameLike
          in: query
          description: result.nameのlike検索
          required: false
          explode: false
          schema:
            type: string
        - name: resultId
          in: query
          description: result.idの検索
          required: false
          explode: false
          schema:
            type: string
        - name: materialItemIds
          in: query
          description: material.item.idの検索（カンマつなぎで複数指定可）
          required: false
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: "成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeArray"
              examples:
                RecipeArray:
                  $ref: "#/components/examples/RecipeArrayExample"
    post:
      tags:
        - recipe
      summary: Recipeの新規作成
      description: ""
      operationId: createRecipe
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                resultId:
                  type: integer
                amount:
                  type: integer
                  minimum: 1
                materials:
                  type: array
                  items:
                    type: object
                    properties:
                      itemId:
                        type: integer
                      amount:
                        type: integer
                        minimum: 1
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                  resultId:
                    type: integer
                  amount:
                    type: integer
        400:
          description: fail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        500:
          description: fail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        name:
          type: string
        tradeInPrice:
          type: integer
        size:
          type: integer
    ItemArray:
      type: array
      items:
        $ref: "#/components/schemas/Item"
    Material:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        recipeId:
          type: integer
        itemId:
          type: integer
        amount:
          type: integer
        item:
          $ref: "#/components/schemas/Item"
    MaterialArray:
      type: array
      items:
        $ref: "#/components/schemas/Material"
    Recipe:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        resultId:
          type: integer
        amount:
          type: integer
        result:
          $ref: "#/components/schemas/Item"
        materials:
          $ref: "#/components/schemas/MaterialArray"
    RecipeArray:
      type: array
      items:
        $ref: "#/components/schemas/Recipe"
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
  examples:
    ItemArrayExample:
      value:
        - id: 1
          createdAt: "2025-02-08T00:00:00.000Z"
          updatedAt: "2025-02-08T00:00:00.000Z"
          name: "きのこの天ぷら"
          price: 130
          tradeInPrice: 65
          healAmount: 15
          healSecond: 4
          size: 2
        - id: 2
          createdAt: "2025-02-08T00:00:00.000Z"
          updatedAt: "2025-02-08T00:00:00.000Z"
          name: "ベルベットの鐘"
          price: 10
          tradeInPrice: 5
          healAmount: 1
          healSecond: 1
          size: 1
    MaterialArrayExample:
      value:
        - id: 1
          createdAt: "2025-02-08T00:00:00.000Z"
          updatedAt: "2025-02-08T00:00:00.000Z"
          recipeId: 1
          itemId: 2
          amount: 8
          item:
            id: 2
            createdAt: "2025-02-08T00:00:00.000Z"
            updatedAt: "2025-02-08T00:00:00.000Z"
            name: "ベルベットの鐘"
            price: 10
            tradeInPrice: 5
            healAmount: 1
            healSecond: 1
            size: 1
    RecipeArrayExample:
      value:
        - id: 1
          createdAt: "2025-02-08T00:00:00.000Z"
          updatedAt: "2025-02-08T00:00:00.000Z"
          recipeId: 1
          itemId: 2
          amount: 8
          result:
            - id: 1
              createdAt: "2025-02-08T00:00:00.000Z"
              updatedAt: "2025-02-08T00:00:00.000Z"
              name: "きのこの天ぷら"
              tradeInPrice: 65
              healAmount: 15
              healSecond: 4
              size: 2
          material:
            - id: 1
              createdAt: "2025-02-08T00:00:00.000Z"
              updatedAt: "2025-02-08T00:00:00.000Z"
              recipeId: 1
              itemId: 2
              amount: 8
              item:
                id: 2
                createdAt: "2025-02-08T00:00:00.000Z"
                updatedAt: "2025-02-08T00:00:00.000Z"
                name: "ベルベットの鐘"
                price: 10
                tradeInPrice: 5
                healAmount: 1
                healSecond: 1
                size: 1